requirements.txt
pytorch-lightning==1.6.1
sacrebleu
jsonlines
hydra-core
transformers==4.18.0
datasets==1.17.0
tqdm
nltk
rouge_score
lyricsgenius
pronouncing
phonemizer
wordfreq
wandb
sentencepiece
pyyaml==5.4.1
# git+https://github.com/PytorchLightning/lightning-bolts.git@master
pypinyin
hydra
torch==1.8.1